Доделана основная часть, осталось часть проверок и ответы в нужном формате

+++ ewm-service/service/src/main/java/ru/practicum/pub/compilations/CompilationsController.java
@@ -26,7 +26,7 @@ public class CompilationsController {
     }
 
     @GetMapping("/{compId}")
-    public CompilationDto getCompilation(@PathVariable Integer compId) {
+    public CompilationDto getCompilation(@PathVariable long compId) {
         log.info("GET /compilations/{}", compId);
         return compilationsService.getCompilation(compId);
     }
